+++ diamondback/filters/complex_frequency_filter.py
@@ -58,7 +58,7 @@
         Larry Turner, Schneider Electric, AI Hub, 2018-02-01.
 """
 
-from diamondback.filters.FirFilter import FirFilter
+from diamondback.filters.fir_filter import FirFilter
 from typing import Tuple, Union
 import math
 import numpy
