+++ diamondback/filters/integral_filter.py
@@ -74,7 +74,7 @@
         Larry Turner, Schneider Electric, AI Hub, 2018-02-06.
 """
 
-from diamondback.filters.IirFilter import IirFilter
+from diamondback.filters.iir_filter import IirFilter
 from typing import Union
 import numpy
