+++ diamondback/filters/pid_filter.py
@@ -35,7 +35,7 @@
         Larry Turner, Schneider Electric, AI Hub, 2018-01-31.
 """
 
-from diamondback.filters.FirFilter import FirFilter
+from diamondback.filters.fir_filter import FirFilter
 from typing import Union
 import numpy
